requirements.txt
Pillow==9.1.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
